requirements.txt
google-adk==1.1.1
mcp==1.9.2
fastapi==0.115.0
uvicorn==0.34.0
python-multipart==0.0.9
langchain==0.3.25
pydantic==2.11.5
pypdf==4.1.0
google-generativeai==0.3.2
numpy==2.2.6
